Refactored PortalDb.
