+++ backend/tests/Logitar.Portal.IntegrationTests/Application/Dictionaries/Commands/ReplaceDictionaryCommandTests.cs
@@ -5,6 +5,7 @@ using Logitar.Portal.Domain.Dictionaries;
 using Logitar.Portal.EntityFrameworkCore.Relational;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
+using PortalDb = Logitar.Portal.EntityFrameworkCore.Relational.PortalDb;
 
 namespace Logitar.Portal.Application.Dictionaries.Commands;
