+++ backend/tests/Logitar.Portal.IntegrationTests/Application/Messages/Commands/SendMessageInternalCommandTests.cs
@@ -24,6 +24,7 @@ using Logitar.Portal.EntityFrameworkCore.Relational;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using PortalDb = Logitar.Portal.EntityFrameworkCore.Relational.PortalDb;
 
 namespace Logitar.Portal.Application.Messages.Commands;
