+++ backend/tests/Logitar.Portal.IntegrationTests/Application/Messages/Queries/ReadMessageQueryTests.cs
@@ -9,6 +9,7 @@ using Logitar.Portal.Domain.Templates;
 using Logitar.Portal.EntityFrameworkCore.Relational;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
+using PortalDb = Logitar.Portal.EntityFrameworkCore.Relational.PortalDb;
 
 namespace Logitar.Portal.Application.Messages.Queries;
