+++ backend/tests/Logitar.Portal.IntegrationTests/Application/Realms/Commands/ReplaceRealmCommandTests.cs
@@ -5,6 +5,7 @@ using Logitar.Portal.Domain.Realms;
 using Logitar.Portal.EntityFrameworkCore.Relational;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
+using PortalDb = Logitar.Portal.EntityFrameworkCore.Relational.PortalDb;
 
 namespace Logitar.Portal.Application.Realms.Commands;
