+++ backend/tests/Logitar.Portal.IntegrationTests/Application/Senders/Commands/SetDefaultSenderCommandTests.cs
@@ -7,6 +7,7 @@ using Logitar.Portal.Domain.Senders.SendGrid;
 using Logitar.Portal.EntityFrameworkCore.Relational;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
+using PortalDb = Logitar.Portal.EntityFrameworkCore.Relational.PortalDb;
 
 namespace Logitar.Portal.Application.Senders.Commands;
