+++ backend/tests/Logitar.Portal.IntegrationTests/Application/Senders/Queries/SearchSendersQueryTests.cs
@@ -8,6 +8,7 @@ using Logitar.Portal.Domain.Senders.SendGrid;
 using Logitar.Portal.EntityFrameworkCore.Relational;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
+using PortalDb = Logitar.Portal.EntityFrameworkCore.Relational.PortalDb;
 
 namespace Logitar.Portal.Application.Senders.Queries;
