+++ backend/tests/Logitar.Portal.IntegrationTests/Application/Templates/Commands/ReplaceTemplateCommandTests.cs
@@ -1,5 +1,4 @@
 ﻿using Logitar.Data;
-using Logitar.Data.SqlServer;
 using Logitar.Identity.Domain.Shared;
 using Logitar.Portal.Application.Realms;
 using Logitar.Portal.Contracts.Templates;
@@ -7,6 +6,7 @@ using Logitar.Portal.Domain.Templates;
 using Logitar.Portal.EntityFrameworkCore.Relational;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
+using PortalDb = Logitar.Portal.EntityFrameworkCore.Relational.PortalDb;
 
 namespace Logitar.Portal.Application.Templates.Commands;
